refactor: consolidate all error handling into src/error/
Move all error-related code to dedicated src/error/ directory:

src/error/
├── apiErrorHandler.ts      # API error handling
├── errorTypes.ts           # Custom error classes
├── errorMessages.ts        # Error message constants
├── GlobalErrorBoundary.tsx # React error boundary (from components/)
├── ServerErrorPage.tsx     # 500 error page (from pages/)
├── NotFoundPage.tsx        # 404 error page (from pages/)
└── index.ts                # Unified barrel export

Benefits:
- All error handling in one place
- Easy to maintain and extend
- Clear separation of concerns
- Unified import: import { GlobalErrorBoundary, handleApiError } from '@/error'

Path Alias Update:
- Added @/error/* → src/error/*

Build Status: ✅ Successful
