+++ src/error/errorTypes.ts
