+++ src/error/index.ts
@@ -2,10 +2,15 @@
  * 파일명: index.ts
  * 
  * 파일 용도:
- * 에러 처리 모듈 barrel export
+ * 에러 처리 모듈 통합 barrel export
+ * - 에러 핸들링 유틸리티
+ * - 에러 타입 및 클래스
+ * - 에러 컴포넌트 (ErrorBoundary, 에러 페이지)
  */
 
+// ============================================
 // API 에러 핸들러
+// ============================================
 export {
   handleApiError,
   getErrorMessage,
@@ -17,7 +22,9 @@ export {
 } from './apiErrorHandler';
 export type { ApiErrorResponse, HandledError } from './apiErrorHandler';
 
-// 에러 타입
+// ============================================
+// 에러 타입 및 클래스
+// ============================================
 export {
   AppError,
   ValidationError,
@@ -35,7 +42,9 @@ export {
   isServerError,
 } from './errorTypes';
 
-// 에러 메시지
+// ============================================
+// 에러 메시지 상수
+// ============================================
 export {
   HTTP_ERROR_MESSAGES,
   NETWORK_ERROR_MESSAGES,
@@ -46,3 +55,9 @@ export {
   getHttpErrorMessage,
 } from './errorMessages';
 
+// ============================================
+// 에러 컴포넌트
+// ============================================
+export { GlobalErrorBoundary } from './GlobalErrorBoundary';
+export { ServerErrorPage } from './ServerErrorPage';
+export { NotFoundPage } from './NotFoundPage';
