+++ tsconfig.app.json
@@ -25,6 +25,7 @@
       "@/hooks/*": ["src/hooks/*"],
       "@/stores/*": ["src/stores/*"],
       "@/common/*": ["src/common/*"],
+      "@/error/*": ["src/error/*"],
       "@/types/*": ["src/types/*"]
     },
